requirements.txt
# requirements.txt

pandas
datasets
torch
sentence-transformers
faiss-cpu
openai
streamlit
